backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
bcrypt
pydantic
scikit-surprise
numpy<2.0.0